[meetups] - Swapped out MFAViews as Updated is now testable - RAG
