+++ ExampleApps/SwiftUIExample/Views/UpdatedViews/MFAView.swift
@@ -1,5 +1,5 @@
 //
-//  MFAuthenticationView.swift
+//  MFAView.swift
 //  SwiftUIExample
 //
 //  Created by Tyler Thompson on 7/15/21.
@@ -13,6 +13,7 @@ struct MFAView: View, PassthroughFlowRepresentable {
     @State var pushSent = false
     @State var enteredCode = ""
     @State var errorMessage: ErrorMessage?
+    @State private var id = UUID()
 
     let inspection = Inspection<Self>() // ViewInspector
     weak var _workflowPointer: AnyFlowRepresentable?
@@ -21,18 +22,19 @@ struct MFAView: View, PassthroughFlowRepresentable {
         VStack(spacing: 30) {
             if !pushSent {
                 Text("This is your friendly MFA Assistant! Tap the button below to pretend to send a push notification and require an account code")
-                Button {
+                PrimaryButton(title: "Start MFA") {
                     withAnimation { pushSent = true }
-                } label: {
-                    Text("Start MFA")
-                        .primaryButtonStyle()
                 }
             } else {
-                Text("Code (enter 1234 to proceed): ").font(.title)
-                TextField("Enter Code:", text: $enteredCode)
-                Button("Submit") {
+                Text("Code (enter 1234 to proceed)").font(.title)
+
+                PrimaryTextField(label: "Code", placeholder: "Enter Code", image: Image(systemName: "number"), text: $enteredCode)
+
+                PrimaryButton(title: "Submit") {
                     if enteredCode == "1234" {
-                        proceedInWorkflow()
+                        withAnimation {
+                            proceedInWorkflow()
+                        }
                     } else {
                         errorMessage = ErrorMessage(message: "Invalid code entered, abandoning workflow.")
                     }
@@ -60,7 +62,7 @@ extension MFAView {
     }
 }
 
-struct MFAuthenticationView_Previews: PreviewProvider {
+struct MFAView_Previews: PreviewProvider {
     static var previews: some View {
         MFAView(with: .none)
             .preferredColorScheme(.dark)
